Move SqlConn to sdk
