+++ db/src/main/scala/squeryl/api/BlockBondsDbApiImpl.scala
@@ -4,10 +4,11 @@ import cats.Applicative
 import cats.syntax.all.*
 import sdk.api.BlockBondsDbApi
 import sdk.api.data.BlockBonds
+import sdk.db.SqlConn
 import squeryl.RhonixNodeDb.blockBondsTable
 import squeryl.tables.BlockBondsTable
 import squeryl.CustomTypeMode.*
-import squeryl.{withSession, SqlConn}
+import squeryl.withSession
 
 class BlockBondsDbApiImpl[F[_]: Applicative: SqlConn] extends BlockBondsDbApi[F] {
   override def insert(blockBonds: BlockBonds): F[BlockBonds] =
